// front-end/app/user-interface/adapter/animation/types.ts
export type BaseZoneName =
  | 'deck'                // 牌堆
  | 'field'               // 場牌區
  | 'player-hand'         // 玩家手牌
  | 'opponent-hand'       // 對手手牌
  | 'player-depository'   // 玩家獲得區
  | 'opponent-depository' // 對手獲得區

/**
 * 獲得區分組區域名稱
 */
export type DepositoryGroupZoneName =
  | `player-depository-${CardType}`
  | `opponent-depository-${CardType}`

/**
 * 區域名稱聯合型別
 *
 * 包含基礎區域和獲得區分組區域
 */
export type ZoneName = BaseZoneName | DepositoryGroupZoneName